Add files via upload
Object recognition is a computer vision technique for identifying objects in image or videos.
Object recognition is a key output of deep learning and machine learning algorithms. There
are a few ways in which artificial intelligence can be used to help those who are blind or
have low vision so in this way we can help poor people. Since the cost of the device is budget
friendly and easily usable it can help a lot of people